requirements.txt
#psycopg2==2.9.3
psycopg2-binary
redis==4.3.3
Requests==2.31.0
rq==1.10.1
pandas
lxml
html5lib
beautifulsoup4